Reorganize controllers

+++ app/Http/Controllers/Part/PartReleaseController.php
@@ -1,50 +1,24 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Part;
 
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Bus;
 
+use App\Http\Controllers\Controller;
+use App\Http\Requests\PartReleaseCreateStep1Request;
+use App\Http\Requests\PartReleaseCreateStep2Request;
 use App\Models\Part;
-use App\Models\PartRelease;
 
 use App\LDraw\LDrawFileValidate;
 use App\LDraw\LibraryOperations;
 
-class ReleaseController extends Controller
+class PartReleaseController extends Controller
 {
-  public function create(Request $request, $step = null) {
+  protected function create() {
     $this->authorize('create', PartRelease::class);
-    set_time_limit(0);
-    switch($step) {
-      case 3:
-        $validated = $request->validate([
-          'ids.*' => 'required|integer',
-          'approve' => 'required:accepted',
-        ]);
-        set_time_limit(30);
-        return $this->doStep3($validated['ids'], $request->file('ldrawfile'));
-        break;
-      case 2:
-        $validated = $request->validate([
-          'ldrawfiles.*' => 'sometimes|file',
-          'ids.*' => 'required|integer',
-        ]);
-        set_time_limit(30);
-        return $this->doStep2($validated['ids'], $validated['ldrawfiles'] ?? []);
-        break;
-      case 1:
-      default:
-        set_time_limit(30);
-        return $this->doStep1();
-        break;
-    }
-  }
-    
-  protected function doStep1() {
-    $parts = Part::unofficial()->where('vote_sort', 1)->orderBy('part_type_id')->get();
+    $parts = Part::unofficial()->where('vote_sort', 1)->orderBy('part_type_id')->orderBy('filename')->get();
     $results = [];
     foreach($parts as $part) {
       $text = $part->isTexmap() ? $part->header : $part->get();
@@ -67,37 +41,42 @@ class ReleaseController extends Controller
       
       $results[] = ['part' => $part, 'release' => empty($errors) && $part->releasable(), 'errors' => $errors, 'warnings' => $warnings];
     }
-    return view('tracker.release.create.step1', ['parts' => $results]);
+    return view('tracker.release.create', ['parts' => $results]);
   }
 
-  protected function doStep2($ids, $ldrawfiles = []) {
+  protected function createStep2(PartReleaseCreateStep1Request $request) {
+    $this->authorize('create', PartRelease::class);
+    $data = $request->validated();
     $filelist = [];
     $sdisk = config('ldraw.staging_dir.disk');
     $spath = config('ldraw.staging_dir.path');
     if (Storage::disk($sdisk)->exists("$spath/ldraw"))
       Storage::disk($sdisk)->deleteDirectory("$spath/ldraw");
-    foreach ($ldrawfiles as $file) {
+    foreach ($data['ldrawfiles'] ?? [] as $file) {
       $filename = "$spath/ldraw/" . $file->getClientOriginalName();
       Storage::disk($sdisk)->put($filename, $file->get());
       $filelist[] = Storage::disk($sdisk)->url($filename);
     }
-    return view('tracker.release.create.step2', ['parts' => Part::whereIn('id', $ids)->lazy(), 'files' => $filelist]);
+    return view('tracker.release.create2', ['parts' => Part::whereIn('id', $data['ids'])->lazy(), 'files' => $filelist]);
   }
 
-  protected function doStep3($ids) {
-    foreach(Part::whereIn('id', $ids)->lazy() as $part) {
+  protected function store(PartReleaseCreateStep2Request $request) {
+    $this->authorize('store', PartRelease::class);
+    $data = $request->validated();
+    foreach(Part::whereIn('id', $data['ids'])->lazy() as $part) {
       LibraryOperations::getAllParentIds($part, $unf_render_list, true);
     }  
-    $unf_render_list = array_diff($unf_render_list, $ids);
-
+    $unf_render_list = array_diff($unf_render_list, $data['ids']);
+/*
     Bus::batch([[
-      new \App\Jobs\Release\MakePartRelease($ids, Auth::user()),
+      new \App\Jobs\Release\MakePartRelease($data['ids'], Auth::user()),
       new \App\Jobs\Release\MakeReleaseZip,
       new \App\Jobs\UpdateSubparts(false),
       new \App\Jobs\UpdateUncertifiedSubparts(true),
       new \App\Jobs\Release\PostReleaseCleanup($unf_render_list),
     ]])->then(function ($batch) {
     })->dispatch();
+*/    
     return redirect()->route('tracker.activity');
   }
